tests - alter table, delete rows refactor
